Create mb example.py
The purpose of [manage branches](https://huggingface.co/Anthonyg5005/hf-scripts/blob/main/manage%20branches.py) is to run this simple command, but it also tries to make it easier because you don't require typing this and editing it every time you want to create or even delete a branch. It also makes sure you have everything configured correctly to run the command.